+++ api/src/app.module.ts
@@ -1,6 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthenticationModule } from 'modules/authentication/authentication.module';
+import { CountriesModule } from 'modules/countries/countries.module';
+import { ScenariosModule } from 'modules/scenarios/scenarios.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { PingController } from './modules/ping/ping.controller';
@@ -10,7 +12,9 @@ import { UsersModule } from './modules/users/users.module';
 @Module({
   imports: [
     TypeOrmModule.forRoot(),
+    CountriesModule,
     ProjectsModule,
+    ScenariosModule,
     UsersModule,
     AuthenticationModule,
   ],