+++ api/CHANGELOG.md
@@ -24,15 +24,14 @@ Unreleased
 - Stub `ProjectsModule`.
 - Stub roles.
 - PoC JSON:API serialization.
-<<<<<<< HEAD
 - Authentication
   - local authentication strategy (username+password)
   - JWT tokens are issued
   - initial support for JWT token invalidation
   - refresh of a still-valid JWT token
-=======
 - Ping/liveness endpoint.
->>>>>>> develop
+- NestJS services are based on `nestjs-base-service`.
+- Integration with faker.js for mock data.
 
 ### Changed